Changed Nextcloud URL regular expression
Nextcloud changed the syntax of their HTML. They were
no longer using quotes around the href attribute, which
is frustratingly valid HTML5. For simplicity, the regex
now no longer looks for an href attribute but only the
desired URI pattern with word-boundary anchors on either side
of it. The version number character class now looks for any
character that is not whitespace, a quote, a slash, or a closing
angle bracket.